Refactor(functions): modularização das funções
