--- Esquema_permissoes
@@ -1,1 +1,0 @@
-1.
